requirements.txt
diffusers==0.32.2
gradio==5.15.0
langchain==0.3.18
langchain_google_genai==2.0.9
Pillow==11.1.0
python-dotenv==1.0.1
transformers==4.46.1
transformers==4.48.3
